requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.2.0
python-dotenv>=1.0.0
